Add methods to User class
